Warnings

+++ include/stp/Simplifier/constantBitP/MersenneTwister.h
@@ -261,7 +261,7 @@ inline void MTRand::seed(uint32* const bigSeed, const uint32 seedLength)
   initialize(19650218UL);
   int i = 1;
   uint32 j = 0;
-  int k = (N > seedLength ? N : seedLength);
+  int k = ((int)N > seedLength ? (int)N : seedLength);
   for (; k; --k)
   {
     state[i] = state[i] ^ ((state[i - 1] ^ (state[i - 1] >> 30)) * 1664525UL);
